Refactor funcs
